Switched to multi-file organization
