+++ functions/src/index.ts
@@ -6,16 +6,6 @@ const app: Express = express()
 
 admin.initializeApp()
 
-const firebaseConfig = {
-  apiKey: 'AIzaSyD1yV--rl-qJiyvwju2K9jz_jkhvr8sTHw',
-  authDomain: 'ofast-e6866.firebaseapp.com',
-  projectId: 'ofast-e6866',
-  storageBucket: 'ofast-e6866.appspot.com',
-  messagingSenderId: '660869453090',
-  appId: '1:660869453090:web:b919fe7e93c35a77a5417b',
-  measurementId: 'G-3B0LRWZFH5',
-}
-
 import cors from 'cors'
 
 const allowedOriginsList = [
@@ -43,9 +33,6 @@ app.use(
   }),
 )
 
-import { initializeApp } from 'firebase/app'
-const appInit = initializeApp(firebaseConfig)
-
 /**
  * Test API
  *
@@ -56,22 +43,8 @@ app.post('/helloWorld', (req: Request, res: Response) => {
   res.json({ str: 'Hello World!' })
 })
 
-import {
-  getAuth,
-  signInWithEmailAndPassword,
-  createUserWithEmailAndPassword,
-  Auth,
-} from 'firebase/auth'
-import {
-  getFirestore,
-  Firestore,
-  doc,
-  setDoc,
-  getDoc,
-} from 'firebase/firestore'
-
-const auth: Auth = getAuth()
-const db: Firestore = getFirestore(appInit)
+import { emailLogin, emailRegister } from './user'
+import { getUserData } from './userData'
 
 /**
  * API for logging in via an email and password
@@ -86,24 +59,7 @@ const db: Firestore = getFirestore(appInit)
  *          - Missing Email: No email field was provided
  *          - Missing Password: No password field was provided
  */
-app.post('/loginWithEmail', (req: Request, res: Response) => {
-  const user = req.body
-  signInWithEmailAndPassword(auth, user.email, user.password)
-    .then((data) => {
-      return res.status(200).json({ userID: data.user.uid })
-    })
-    .catch((err) => {
-      if (err.code === 'auth/invalid-login-credentials')
-        return res.status(401).json({ general: 'Invalid Credentials' })
-      else if (err.code === 'auth/invalid-email')
-        return res.status(401).json({ general: 'Invalid Email' })
-      else if (err.code === 'auth/missing-email')
-        return res.status(401).json({ general: 'Missing Email' })
-      else if (err.code === 'auth/missing-password')
-        return res.status(401).json({ general: 'Missing Password' })
-      else return res.status(500).json({ error: err.code })
-    })
-})
+app.post('/loginWithEmail', emailLogin)
 
 /**
  * API for registering via an email and password
@@ -118,30 +74,7 @@ app.post('/loginWithEmail', (req: Request, res: Response) => {
  *          - Missing Email: No email field was provided
  *          - Missing Password: No password field was provided
  */
-app.post('/registerWithEmail', (req: Request, res: Response) => {
-  const newUser = req.body
-  createUserWithEmailAndPassword(auth, newUser.email, newUser.password)
-    .then((data) => {
-      setDoc(doc(db, 'UserData', data.user.uid), {})
-        .then(() => {
-          return res.status(201).json({ general: 'User Created' })
-        })
-        .catch((err) => {
-          return res.status(500).json({ error: err })
-        })
-    })
-    .catch((err) => {
-      if (err.code === 'auth/email-already-in-use')
-        return res.status(401).json({ general: 'Email in Use' })
-      else if (err.code === 'auth/invalid-email')
-        return res.status(401).json({ general: 'Invalid Email' })
-      else if (err.code === 'auth/missing-email')
-        return res.status(401).json({ general: 'Missing Email' })
-      else if (err.code === 'auth/missing-password')
-        return res.status(401).json({ general: 'Missing Password' })
-      else return res.status(500).json({ error: err.code })
-    })
-})
+app.post('/registerWithEmail', emailRegister)
 
 /**
  * API for retrieving user data
@@ -162,33 +95,6 @@ app.post('/registerWithEmail', (req: Request, res: Response) => {
  *          - problemsTLE
  *          - ProblemsWrong
  */
-app.get('/getUserData', (req: Request, res: Response) => {
-  const userId: string = req.body.uid
-  admin
-    .auth()
-    .getUser(userId)
-    .then((curUser) => {
-      let data = {
-        email: curUser.email,
-        username: curUser.displayName,
-      }
-
-      getDoc(doc(db, 'UserData', userId))
-        .then((doc) => {
-          if (doc.exists()) {
-            data = { ...data, ...doc.data() }
-            return res.status(200).json(data)
-          } else {
-            return res.status(404).json({ general: 'User Data Not Found' })
-          }
-        })
-        .catch((err) => {
-          return res.status(500).json({ error: err.code })
-        })
-    })
-    .catch((err) => {
-      return res.status(500).json({ error: err.code })
-    })
-})
+app.post('/getUserData', getUserData)
 
 exports.api = https.onRequest(app)
